+++ src/rs/symbol.rs
@@ -32,6 +32,9 @@ pub struct Symbol {
     /// set to `false`.
     pub nonmatching_sym_exists: bool,
 
+    /// `true` if this symbol has a `.NON_MATCHING` suffix.
+    pub is_nonmatching: bool,
+
     /// This symbol was not originally present on the mapfile, but instead it
     /// was added during parsing because it was inferred this symbol must exist
     /// due to mismatches on the addresses or sizes of other symbols in this
@@ -47,6 +50,7 @@ impl Symbol {
         vrom: Option<u64>,
         align: Option<u64>,
         nonmatching_sym_exists: bool,
+        is_nonmatching: bool,
         inferred_static: bool,
     ) -> Self {
         Self {
@@ -56,20 +60,23 @@ impl Symbol {
             vrom,
             align,
             nonmatching_sym_exists,
+            is_nonmatching,
             inferred_static,
         }
     }
 
     pub fn new(name: String, vram: u64, size: u64, vrom: Option<u64>, align: Option<u64>) -> Self {
-        Self::new_impl(name, vram, size, vrom, align, false, false)
+        let is_nonmatching = name.ends_with("NON_MATCHING");
+        Self::new_impl(name, vram, size, vrom, align, false, is_nonmatching, false)
     }
 
     pub fn new_default(name: String, vram: u64) -> Self {
-        Self::new_impl(name, vram, 0, None, None, false, false)
+        let is_nonmatching = name.ends_with("NON_MATCHING");
+        Self::new_impl(name, vram, 0, None, None, false, is_nonmatching, false)
     }
 
     pub fn new_static(name: String, vram: u64, size: u64, vrom: Option<u64>) -> Self {
-        Self::new_impl(name, vram, size, vrom, None, false, true)
+        Self::new_impl(name, vram, size, vrom, None, false, false, true)
     }
 
     pub fn get_vram_str(&self) -> String {
@@ -140,7 +147,7 @@ pub(crate) mod python_bindings {
     #[pymethods]
     impl super::Symbol {
         #[new]
-        #[pyo3(signature=(name,vram,size=0,vrom=None,align=None, nonmatchingSymExists=false, inferredStatic=false))]
+        #[pyo3(signature=(name,vram,size=0,vrom=None,align=None, nonmatchingSymExists=false, isNonmatching=false, inferredStatic=false))]
         fn py_new(
             name: String,
             vram: u64,
@@ -148,6 +155,7 @@ pub(crate) mod python_bindings {
             vrom: Option<u64>,
             align: Option<u64>,
             nonmatchingSymExists: bool,
+            isNonmatching: bool,
             inferredStatic: bool,
         ) -> Self {
             Self::new_impl(
@@ -157,6 +165,7 @@ pub(crate) mod python_bindings {
                 vrom,
                 align,
                 nonmatchingSymExists,
+                isNonmatching,
                 inferredStatic,
             )
         }
@@ -226,6 +235,17 @@ pub(crate) mod python_bindings {
         #[setter]
         fn set_nonmatchingSymExists(&mut self, value: bool) -> PyResult<()> {
             self.nonmatching_sym_exists = value;
+            Ok(())
+        }
+
+        #[getter]
+        fn get_isNonmatching(&self) -> PyResult<bool> {
+            Ok(self.is_nonmatching)
+        }
+
+        #[setter]
+        fn set_isNonmatching(&mut self, value: bool) -> PyResult<()> {
+            self.is_nonmatching = value;
             Ok(())
         }