+++ tests/update_outputs.py
@@ -33,12 +33,14 @@ prefixesToTrim.append("build/")
 reportCategories = mapfile_parser.ReportCategories()
 reportCategories.setReportData(True)
 
-def plfResolver(x: Path) -> Path|None:
+
+def plfResolver(x: Path) -> Path | None:
     if x.suffix == ".plf":
         newPath = x.with_suffix(".map")
         if newPath.exists():
             return newPath
     return None
+
 
 for mapPath in sorted(mapsPath.rglob("*")):
     if not mapPath.is_file():
@@ -47,18 +49,32 @@ for mapPath in sorted(mapsPath.rglob("*")):
     print(mapPath)
 
     print("    .json")
-    mapfile_parser.frontends.jsonify.doJsonify(mapPath, outputFolder/mapPath.with_suffix(".json").relative_to(mapsPath), plfResolver=plfResolver)
+    mapfile_parser.frontends.jsonify.doJsonify(
+        mapPath,
+        outputFolder / mapPath.with_suffix(".json").relative_to(mapsPath),
+        plfResolver=plfResolver,
+    )
 
     print("    .machine.json")
-    mapfile_parser.frontends.jsonify.doJsonify(mapPath, outputFolder/mapPath.with_suffix(".machine.json").relative_to(mapsPath), humanReadable=False, plfResolver=plfResolver)
+    mapfile_parser.frontends.jsonify.doJsonify(
+        mapPath,
+        outputFolder / mapPath.with_suffix(".machine.json").relative_to(mapsPath),
+        humanReadable=False,
+        plfResolver=plfResolver,
+    )
 
     print("    .sym")
-    mapfile_parser.frontends.pj64_syms.doPj64Syms(mapPath, outputFolder/mapPath.with_suffix(".sym").relative_to(mapsPath), plfResolver=plfResolver)
+    mapfile_parser.frontends.pj64_syms.doPj64Syms(
+        mapPath,
+        outputFolder / mapPath.with_suffix(".sym").relative_to(mapsPath),
+        plfResolver=plfResolver,
+    )
 
     print("    .objdiff_report.json")
     mapfile_parser.frontends.objdiff_report.doObjdiffReport(
         mapPath,
-        outputFolder/mapPath.with_suffix(".objdiff_report.json").relative_to(mapsPath),
+        outputFolder
+        / mapPath.with_suffix(".objdiff_report.json").relative_to(mapsPath),
         prefixesToTrim,
         reportCategories,
         quiet=True,
@@ -66,7 +82,16 @@ for mapPath in sorted(mapsPath.rglob("*")):
     )
 
     print("    .csv")
-    mapfile_parser.frontends.symbol_sizes_csv.doSymbolSizesCsv(mapPath, outputFolder/mapPath.with_suffix(".csv").relative_to(mapsPath), plfResolver=plfResolver)
+    mapfile_parser.frontends.symbol_sizes_csv.doSymbolSizesCsv(
+        mapPath,
+        outputFolder / mapPath.with_suffix(".csv").relative_to(mapsPath),
+        plfResolver=plfResolver,
+    )
 
     print("    .symbols.csv")
-    mapfile_parser.frontends.symbol_sizes_csv.doSymbolSizesCsv(mapPath, outputFolder/mapPath.with_suffix(".symbols.csv").relative_to(mapsPath), symbolsSummary=True, plfResolver=plfResolver)
+    mapfile_parser.frontends.symbol_sizes_csv.doSymbolSizesCsv(
+        mapPath,
+        outputFolder / mapPath.with_suffix(".symbols.csv").relative_to(mapsPath),
+        symbolsSummary=True,
+        plfResolver=plfResolver,
+    )